price is float

+++ db/migrate/20150212203014_create_prices.rb
@@ -1,8 +1,8 @@
 class CreatePrices < ActiveRecord::Migration
   def change
     create_table :prices do |t|
-      t.integer :trefis_analyst_price
-      t.integer :market_price
+      t.float :trefis_analyst_price
+      t.float :market_price
       t.integer :stock_id
       t.date    :date
       t.timestamps
